requirements.txt
# FastAPI and Web
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
python-multipart>=0.0.9,<1.0.0

# Database
sqlalchemy[asyncio]>=2.0.0,<3.0.0
asyncpg>=0.29.0,<1.0.0
alembic>=1.13.0,<2.0.0

# Redis
redis>=5.0.0,<6.0.0

# Authentication
PyJWT>=2.8.0,<3.0.0
passlib[bcrypt]>=1.7.4,<2.0.0

# Configuration
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
email-validator>=2.0.0,<3.0.0

# Logging
structlog>=24.0.0,<25.0.0

# Testing
pytest>=8.2.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
httpx>=0.26.0,<1.0.0

# Development
python-dotenv>=1.0.0,<2.0.0
